// marketplace/src/types/custom-interface.ts
export interface IleaveMessageFormValues {
  name: string;
  email: string;
  phone: string;
  message: string;
}